Removed ScreamException.
