+++ src/test/java/de/michab/scream/fcos/BytevectorTest.java
@@ -14,8 +14,8 @@ import static org.junit.jupiter.api.Assertions.fail;
 import org.junit.jupiter.api.Test;
 
 import de.michab.scream.RuntimeX;
+import de.michab.scream.RuntimeX.Code;
 import de.michab.scream.ScreamBaseTest;
-import de.michab.scream.ScreamException.Code;
 
 public class BytevectorTest extends ScreamBaseTest
 {
