+++ src/test/java/de/michab/scream/fcos/EnvironmentTest.java
@@ -13,8 +13,8 @@ import static org.junit.jupiter.api.Assertions.fail;
 import org.junit.jupiter.api.Test;
 
 import de.michab.scream.RuntimeX;
+import de.michab.scream.RuntimeX.Code;
 import de.michab.scream.ScreamBaseTest;
-import de.michab.scream.ScreamException.Code;
 
 public class EnvironmentTest extends ScreamBaseTest
 {
