+++ src/test/java/de/michab/scream/fcos/SchemeDoubleTest.java
@@ -14,9 +14,8 @@ import static org.junit.jupiter.api.Assertions.fail;
 import org.junit.jupiter.api.Test;
 
 import de.michab.scream.RuntimeX;
+import de.michab.scream.RuntimeX.Code;
 import de.michab.scream.ScreamBaseTest;
-import de.michab.scream.ScreamException;
-import de.michab.scream.ScreamException.Code;
 
 public class SchemeDoubleTest extends ScreamBaseTest
 {
@@ -72,7 +71,7 @@ public class SchemeDoubleTest extends ScreamBaseTest
             throws X;
     }
 
-    private void typeFailureTest( FuncX<FirstClassObject, Number, Exception> op ) throws RuntimeX
+    private void typeFailureTest( FuncX<FirstClassObject, Number, Exception> op ) throws Exception
     {
         var symbol = Symbol.createObject( "313" );
         try
@@ -80,22 +79,18 @@ public class SchemeDoubleTest extends ScreamBaseTest
             op.apply( symbol );
             fail();
         }
-        catch ( Exception e )
+        catch ( RuntimeX rx )
         {
-            assertInstanceOf( ScreamException.class, e );
-            ScreamException se = (ScreamException)e;
-            assertEquals( Code.TYPE_ERROR, se.getCode() );
+            assertEquals( Code.TYPE_ERROR, rx.getCode() );
         }
         try
         {
             op.apply( Cons.NIL );
             fail();
         }
-        catch ( Exception e )
+        catch ( RuntimeX rx )
         {
-            assertInstanceOf( ScreamException.class, e );
-            ScreamException se = (ScreamException)e;
-            assertEquals( Code.TYPE_ERROR, se.getCode() );
+            assertEquals( Code.TYPE_ERROR, rx.getCode() );
         }
     }
