+++ src/test/java/de/michab/scream/language/R7rs_4_1_4_Procedures_Test.java
@@ -9,8 +9,8 @@ import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 
 import org.junit.jupiter.api.Test;
 
+import de.michab.scream.RuntimeX.Code;
 import de.michab.scream.ScreamBaseTest;
-import de.michab.scream.ScreamException.Code;
 import de.michab.scream.fcos.Operation;
 
 /**
