+++ src/test/java/de/michab/scream/language/R7rs_5_3_1_Top_level_definitions_Test.java
@@ -11,8 +11,8 @@ import static org.junit.jupiter.api.Assertions.fail;
 import org.junit.jupiter.api.Test;
 
 import de.michab.scream.RuntimeX;
+import de.michab.scream.RuntimeX.Code;
 import de.michab.scream.ScreamBaseTest;
-import de.michab.scream.ScreamException.Code;
 
 /**
  * r7rs 5.3.1 Top level definitions
