+++ src/test/java/de/michab/scream/language/R7rs_6_12_Environments_and_evaluation_Test.java
@@ -14,8 +14,8 @@ import static org.junit.jupiter.api.Assertions.fail;
 import org.junit.jupiter.api.Test;
 
 import de.michab.scream.RuntimeX;
+import de.michab.scream.RuntimeX.Code;
 import de.michab.scream.ScreamBaseTest;
-import de.michab.scream.ScreamException.Code;
 import de.michab.scream.fcos.Environment;
 
 /**
