+++ src/test/java/de/michab/scream/language/R7rs_6_13_1_Ports_Test.java
@@ -21,8 +21,8 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import de.michab.scream.RuntimeX;
+import de.michab.scream.RuntimeX.Code;
 import de.michab.scream.ScreamBaseTest;
-import de.michab.scream.ScreamException.Code;
 import de.michab.scream.fcos.Cons;
 import de.michab.scream.fcos.Procedure;
 import de.michab.scream.fcos.SchemeBoolean;
