+++ src/test/java/de/michab/scream/pops/SyntaxDoTest.java
@@ -9,9 +9,9 @@ import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 
 import org.junit.jupiter.api.Test;
 
+import de.michab.scream.RuntimeX.Code;
 import de.michab.scream.ScreamBaseTest;
 import de.michab.scream.ScreamEvaluator;
-import de.michab.scream.ScreamException.Code;
 import de.michab.scream.fcos.Operation;
 
 public class SyntaxDoTest extends ScreamBaseTest
